power: supply: qti_battery_charger: limit the ADSP workaround to user limits
The workaround may involve displaying disconnect and reconnect animations
to the user because the system detects a power-cycle, even if the user
didn't change the limit because the stock vendor daemon may constantly
change it.

While the ADSP bug itself is present regardless of who (stock vendor
daemon or the user) writes it, limit it to user limits only and maintain
the stock behavior if the FCC isn't manually touched.

Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>